% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasEquiv_EffectSize_Base.R
\name{colSD}
\alias{colSD}
\title{Standard deviations of columns}
\usage{
colSD(x, ...)
}
\arguments{
\item{x}{is a matrix or data frame for which we want to obtain column sds
\@param ... are other arguments to be passed to \code{sd}, such as
\code{na.rm}}
}
\value{
A vector of standard deviations by column
}
\description{
\code{colSD} computes standard deviations of columns.
}
\keyword{internal}
