% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasEquiv_EffectSize_Base.R
\name{DTF_graph}
\alias{DTF_graph}
\title{Expected bias in test score}
\usage{
DTF_graph(LambdaR, LambdaF, NuR, NuF, RefGroup, FocGroup, ThreshR = NULL,
  ThreshF = NULL, ThetaR = NULL, ThetaF = NULL,
  categorical = FALSE, minEta = -3, maxEta = 3,
  GraphType = "Scores")
}
\arguments{
\item{LambdaR}{is the factor loading of the indicator onto the factor of
interest for the reference group.}

\item{LambdaF}{is the factor loading of the indicator onto the factor of
interest for the focal group.}

\item{NuR}{is the indicator intercept for the reference group.}

\item{NuF}{is the indicator intercept for the focal group.}

\item{ThreshR}{is a vector of thresholds (for categorical indicators)
for the reference group. Defaults to \code{NULL} for continuous
indicators.}

\item{ThreshF}{is a vector of thresholds (for categorical indicators)
for the focal group. Defaults to \code{NULL} for continuous
indicators.}

\item{ThetaR}{is the indicator residual variance in the
reference group. Defaults to \code{NULL} for continuous
indicators.}

\item{ThetaF}{is the indicator residual variance in the
focal group. Defaults to \code{NULL} for continuous
indicators.}

\item{categorical}{is a Boolean variable declaring whether the variables
in the model are ordered categorical. Models in which some variables are
categorical and others are continuous are not supported. If no value is
provided, categorical defaults to \code{FALSE}, although if a vector of
thresholds are provided, categorical will be forced to
\code{TRUE}. A graded response model with probit link (e.g., DWLS in
lavaan or WLSMV in Mplus) is used for categorical variables. If you desire
for other categorical models (e.g., IRT parameterization) to be supported,
e-mail the maintainer.}

\item{minEta}{is the minimum factor score to be considered. Defaults to -3.}

\item{maxEta}{is the maximum factor score to be considered. Defaults to -3.}

\item{MeanF}{is the factor mean in the focal group}

\item{VarF}{is the factor variances in the focal group.}
}
\value{
A graph of expected total score or bias in expected total score due to
measurement nonequivalence, per Nye & Drasgow (2011).
}
\description{
\code{DTF_graph} produces a graph of expected total score or of expected
bias in total score based on item parameters in the reference group and
based on item parameters in the focal group.
}
\details{
\code{DTF_graph} is called by \code{dmacs_summary}, which
in turn is called by \code{\link{lavaan_dmacs}} and
\code{\link{mplus_dmacs}}, which are the only functions in this
package intended for casual users
}
\section{References}{

Nye, C. & Drasgow, F. (2011). Effect size indices for analyses of
measurement equivalence: Understanding the practical importance of
differences between groups. \emph{Journal of Applied Psychology, 96}(5),
966-980.
}

\examples{
LambdaF <- matrix(c(1.00, 0.74,  1.14, 0.92), ncol = 1)
LambdaR <- matrix(c(1.00, 0.76,  1.31, 0.98), ncol = 1)
NuF     <- c(0.00, 1.28, -0.82, 0.44)
NuR     <- c(0.00, 0.65, -0.77, 0.47)
DTF_graph(LambdaR, LambdaF, NuR, NuF)

}
