% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasEquiv_EffectSize_Wrappers.R
\name{dmacs_summary_single}
\alias{dmacs_summary_single}
\title{Summary of measurement nonequivalence effects for a single group}
\usage{
dmacs_summary_single(LambdaR, LambdaF, NuR, NuF, MeanF, VarF, SD,
  ThreshR = NULL, ThreshF = NULL, ThetaR = NULL, ThetaF = NULL,
  categorical = FALSE)
}
\arguments{
\item{LambdaR}{is the factor loading of the indicator onto the factor of
interest for the reference group.}

\item{LambdaF}{is the factor loading of the indicator onto the factor of
interest for the focal group.}

\item{NuR}{is the indicator intercept for the reference group.}

\item{NuF}{is the indicator intercept for the focal group.}

\item{MeanF}{is the factor mean in the focal group}

\item{VarF}{is the factor variances in the focal group.}

\item{SD}{is the indicator standard deviations to be used as
the denominator of the dmacs effect size. This will usually either be
pooled standard deviation for the indicator or the standard deviation
for the indicator in the reference group.}

\item{ThreshR}{is a vector of thresholds (for categorical indicators)
for the reference group. Defaults to \code{NULL} for continuous
indicators.}

\item{ThreshF}{is a vector of thresholds (for categorical indicators)
for the focal group. Defaults to \code{NULL} for continuous
indicators.}

\item{ThetaR}{is the indicator residual variance in the
reference group. Defaults to \code{NULL} for continuous
indicators.}

\item{ThetaF}{is the indicator residual variance in the
focal group. Defaults to \code{NULL} for continuous
indicators.}

\item{categorical}{is a Boolean variable declaring whether the variables
in the model are ordered categorical. Models in which some variables are
categorical and others are continuous are not supported. If no value is
provided, categorical defaults to \code{FALSE}, although if multiple
thresholds are provided for an item, categorical will be forced to
\code{TRUE}. A graded response model with probit link (e.g., DWLS in
lavaan or WLSMV in Mplus) is used for categorical variables. If you desire
for other categorical models (e.g., IRT parameterization) to be supported,
e-mail the maintainer.}
}
\value{
A list of measurement nonequivalence effects from Nye and Drasgow
(2011), including dmacs,
expected bias in the mean score by item, expected bias in the mean total
score, and expected bias in the variance of the total score. Expected bias
in the variance of the total score is only supplied for unidimensional
models in the current version of this package
}
\description{
\code{dmacs_summary_single} returns a summary of measurement non-equivalence
effects given parameters for a focal and reference group.
}
\details{
\code{dmacs_summary_single} is called by \code{dmacs_summary}, which
in turn is called by \code{\link{lavaan_dmacs}} and
\code{\link{mplus_dmacs}}, which are the only functions in this
package intended for casual users
}
\section{References}{

Nye, C. & Drasgow, F. (2011). Effect size indices for analyses of
measurement equivalence: Understanding the practical importance of
differences between groups. \emph{Journal of Applied Psychology, 96}(5),
966-980.
}

\examples{
LambdaF <- matrix(c(1.00, 0.74,  1.14, 0.92), ncol = 1)
LambdaR <- matrix(c(1.00, 0.76,  1.31, 0.98), ncol = 1)
NuF     <- c(0.00, 1.28, -0.82, 0.44)
NuR     <- c(0.00, 0.65, -0.77, 0.47)
MeanF   <- 0.21
VarF    <- 1.76
SD      <- c(2.12, 1.85,  1.12, 3.61)
dmacs_summary_single(LambdaR, LambdaF, NuR, NuF, MeanF, VarF, SD)

}
